♻️ [Refactoring]: html-node.test.tsを機能別に分割
- colors、flexbox、htmlと同様の構造に変更
- __tests__ディレクトリを作成
- 機能別に4つのテストファイルに分割:
  - from: HTMLNodeの作成
  - typeguards: 型ガード関数
  - factory: ファクトリーメソッド
  - utils: ユーティリティメソッド
- describeを削除しフラットな構造に変更
- テスト名を日本語に変更
